Add error message if parsing file panics
Damn, this thing is fast if built with `cargo build --release`...
Maybe because the paths and function calls are in the source file
and executed at compile time? and that's why?
Maybe later, there will be 2 options
-> running with paths as command-line arguments
-> compiling it itself with paths as part of the source file!
I will have to test the performance of both options
This currently built with `--release`` executes in under 2s xDDDD
Compared to the Python version... that's insane Jesus Christ