Improve performance of functions counting the number of plays
See #22
<https://github.com/fsktom/rusty-endsong-parser/issues/22#issuecomment-1488570707>

It looks worse and I think I've done this the first time, then went
to creating aspects, dervied Eq for them, and compared those directly
for it to look nicer. But seems like allocating a new aspect every time
is very time-consuming. Comparing Strings directly without re-allocating
anything is the way to go apparently.

For normal single usage it doesn't matter. But I discovered this while
working on plots and there I have to call such a function for every data
point (bc I don't want to lower resolution it's equivalent to every time
the aspect exists in the dataset heh). There it makes the code 6x faster