Add itertools and use '.unique()' for aspect lists
Seems like it's a 2x improvement for '.artists()'.
Using 'cargo bench' from 7ms to about 3.5ms. Biggest bottleneck is
obviously the allocations with cloning the Strings... But I can't get
rid of it, since the UI tab complete list is a Vec<String>...
Maybe I could make it a Vec<&str> but too much work and it's not
necessary imo